requirements.txt
colourlovers
numpy
pycairo
scipy
